requirements.txt
streamlit
duckdb<=1.4.4
pandas
plotly
lifelines
requests
openpyxl
pyarrow
python-dateutil
pytest
